refactor: removed old dashboard and switched to the new one
