+++ cmd/wallet-web/src/components/SidebarPlugin/SidebarLink.vue
@@ -5,52 +5,50 @@
 -->
 
 <template>
-  <li class="md-list-item">
-    <router-link
-      class="md-list-item-router md-list-item-container md-button-clean"
-      v-bind="$attrs"
-      @click="hideSidebar"
-    >
-      <div class="md-list-item-content md-ripple">
-        <slot>
-          <md-icon>{{ link.icon }}</md-icon>
-          <p>{{ link.name }}</p>
-        </slot>
-      </div>
+  <li
+    class="
+      justify-start
+      items-center
+      w-full
+      h-16
+      font-bold
+      hover:bg-gradient-to-r hover:opacity-100 hover:from-neutrals-black
+      opacity-60
+      bar
+      flex flex-row
+    "
+  >
+    <router-link class="flex flex-row justify-start items-center px-10" v-bind="$attrs">
+      <img v-if="icon" :src="require(`@/assets/img/${icon}`)" class="w-8 h-8" />
+      <div v-else class="w-8 h-8" />
+      <span class="ml-4 text-lg text-neutrals-white">{{ heading }}</span>
     </router-link>
   </li>
 </template>
+
 <script>
 export default {
   name: 'SidebarLink',
-  inject: {
-    autoClose: {
-      default: true,
-    },
-  },
   props: {
-    link: {
-      type: [String, Object],
-      default: () => {
-        return {
-          name: '',
-          path: '',
-          icon: '',
-        };
-      },
-    },
-    tag: {
+    heading: {
       type: String,
-      default: 'router-link',
+      required: true,
     },
-  },
-  methods: {
-    hideSidebar() {
-      if (this.autoClose && this.$sidebar && this.$sidebar.showSidebar === true) {
-        this.$sidebar.displaySidebar(false);
-      }
+    icon: {
+      type: String,
+      required: true,
     },
   },
 };
 </script>
-<style></style>
+
+<style scoped>
+.bar:hover:before {
+  position: absolute;
+  content: '';
+  display: block;
+  background-color: theme('colors.primary.purple.hashita');
+  height: theme('spacing.16');
+  width: 4px;
+}
+</style>
