+++ cmd/wallet-web/src/main.js
@@ -14,8 +14,7 @@ import * as webCredentialHandler from 'web-credential-handler';
 import { mapActions, mapGetters } from 'vuex';
 import VueCookies from 'vue-cookies';
 import '@/assets/css/tailwind.css';
-import SideBar from '@/components/SidebarPlugin';
-import SideBarV2 from '@/components/SidebarPluginV2';
+import Sidebar from '@/components/SidebarPlugin';
 import VueMaterial from 'vue-material';
 import 'vue-material/dist/vue-material.css';
 import './assets/scss/material-dashboard.scss';
@@ -30,8 +29,7 @@ Vue.prototype.$polyfill = polyfill;
 Vue.prototype.$webCredentialHandler = webCredentialHandler;
 
 Vue.use(VueRouter);
-Vue.use(SideBar);
-Vue.use(SideBarV2);
+Vue.use(Sidebar);
 Vue.use(VueCookies);
 Vue.$cookies.config('7d');
 Vue.use(VueMaterial);
