+++ cmd/wallet-web/src/router/routes.js
@@ -82,74 +82,6 @@ export default [
       },
     ],
   },
-  // Temporary path for the new dashboard
-  {
-    path: 'v2',
-    name: 'DashboardLayoutV2',
-    component: load('layout/DashboardLayoutV2'),
-    redirect: { name: 'dashboardV2' },
-    children: [
-      {
-        path: 'dashboard',
-        name: 'dashboardV2',
-        component: load('DashboardV2'),
-        meta: { requiresAuth: true },
-      },
-      {
-        path: 'WebWallet',
-        name: 'web-walletV2',
-        component: load('demos/WebWallet'),
-      },
-      {
-        path: 'DIDManagement',
-        name: 'did-managementV2',
-        component: load('DIDManagement'),
-        meta: { requiresAuth: true },
-      },
-      {
-        path: 'connections',
-        name: 'connectionsV2',
-        component: load('demos/Connections'),
-        meta: { requiresAuth: true },
-      },
-      {
-        path: 'relationships',
-        name: 'relationshipsV2',
-        component: load('demos/Relationships'),
-        meta: { requiresAuth: true },
-      },
-      {
-        path: 'issue-credential',
-        name: 'issue-credentialV2',
-        component: load('demos/IssueCredential'),
-        meta: { requiresAuth: true },
-      },
-      {
-        path: 'present-proof',
-        name: 'present-proofV2',
-        component: load('demos/PresentProof'),
-        meta: { requiresAuth: true },
-      },
-      {
-        path: 'worker',
-        name: 'chapi-workerV2',
-        component: load('Worker'),
-        meta: { showNav: false, hideFooter: true },
-      },
-      {
-        path: 'StoreInWallet',
-        name: 'chapi-storeV2',
-        component: load('Store'),
-        meta: { blockNoAuth: true, showNav: false, hideFooter: true },
-      },
-      {
-        path: 'GetFromWallet',
-        name: 'chapi-getV2',
-        component: load('Get'),
-        meta: { blockNoAuth: true, showNav: false, hideFooter: true },
-      },
-    ],
-  },
   {
     path: 'loginhandle',
     name: 'loginhandle',
