+++ test/ui-automation/test/helpers/wallet.js
@@ -103,9 +103,16 @@ async function _getSignUp(email) {
 }
 
 async function _logoutWallet() {
-  const logOutButton = await $("button*=Log Out");
+  const logOutButton = await $("button*=Sign Out");
   await logOutButton.waitForExist();
   await logOutButton.click();
+
+  // wait for logout to complele and go to signup page
+  await browser.waitUntil(async () => {
+    const headingLink = await $("h1*=Sign up.");
+    expect(headingLink).toHaveValue("Sign up.");
+    return true;
+  });
 }
 
 async function _signIn(signedUpUserEmail) {
@@ -115,18 +122,11 @@ async function _signIn(signedUpUserEmail) {
   await browser.waitUntil(async () => {
     const signInButton = await $("button*=Demo Sign-In Partner");
     await signInButton.waitForExist();
-    const handles = await browser.getWindowHandles();
     await signInButton.click();
-    await browser.waitUntil(async () => {
-      const newHandles = await browser.getWindowHandles();
-      if (newHandles.length - handles.length === 1) {
-        await _getThirdPartyLogin(signedUpUserEmail);
-        return true;
-      }
-      return false;
-    });
+    await _getThirdPartyLogin(signedUpUserEmail);
     return true;
   });
+  return true;
 }
 
 async function _changeLocale() {
@@ -141,7 +141,16 @@ async function _changeLocale() {
 }
 
 async function _getThirdPartyLogin(email) {
-  await browser.switchWindow("Login Page");
+  await browser.waitUntil(async () => {
+    try {
+      await browser.switchWindow("Login Page");
+    } catch (err) {
+      console.warn("[warn] switch window to login page : ", err.message);
+      return false;
+    }
+    return true;
+  });
+
   await browser.waitUntil(async () => {
     let emailInput = await $("#email");
     await emailInput.waitForExist();
@@ -177,10 +186,7 @@ async function _checkStoredCredentials() {
 }
 
 async function _saveAnyDID({ method }) {
-  const settingsDiv = await $("img[id='dashboardSettings']");
-  await settingsDiv.click();
-
-  const didManager = await $("a*=DID Management");
+  const didManager = await $("a*=Settings");
   await didManager.waitForExist();
   await didManager.click();
 
@@ -226,10 +232,7 @@ async function _saveAnyDID({ method }) {
 }
 
 async function _createTrustblocDID() {
-  const settingsDiv = await $("img[id='dashboardSettings']");
-  await settingsDiv.click();
-
-  const didManager = await $("a*=DID Management");
+  const didManager = await $("a*=Settings");
   await didManager.waitForExist();
   await didManager.click();
